Clean code, and added useRef for focusing input field after adding newTodo
