+++ src/components/HeaderTodo/index.ts
@@ -1,1 +1,1 @@
-export * from './AddingTodo';
+export * from './HeaderTodo';
