+++ src/types/ErrorsType.ts
@@ -1,0 +1,6 @@
+export enum ErrorsType {
+  EMPTY = 'Title cannot be empty',
+  ADD = 'Unable to add todo',
+  DELETE = 'Unable to delete todo',
+  UPDATE = 'Unable to update todo',
+}
